Avoid `.present?` and `.blank?`
`.present?` and `.blank?` are not in plain Ruby; they're extensions
provided by Rails.  Worldwide (is supposed to) support runinng without
Rails, so we need to avoid any Rails-only features.

We already have `Util.blank?` and `Util.present?` that provide
drop-in replacements for this functionality; it's just a matter
of calling our Util method instead.

Our current unit testing runs in Rails, which is why we didn't
spot this at test time.  We should probably revisit that... but
I think that's a broader question for another day.