Patch some territory names (BQ, NL, TF, TR)
There are some names in the CLDR data that don't seem quite right.

BQ (Bonaire, Sint Eustatius and Saba) was formerly part of the
Netherlands Antilles (AN), but AN was dissolved in 2010.
BQ may be referred to as the Caribbean Netherlands.

NL is the country of the Netherlands, which includes (North and South)
Holland as two of its many provinces.  We should make sure that the
country name reflects the fact that we're referring to the country,
not just to a province.

TF's names in Danish and German is a bit too long when compared
to other country names, so we are shortening it.

TR has changed its official English-language name, as recognized
by the UN.

(Update CHANGELOG.md)