Patch data related to 830 (Channel Islands)
This change fixes two problems.

First, the country code for Guernsey is GG.
GY is the postcode area for Guernsey, but country code GY is Guyana.

Second, the designation of numeric-three code 830 for the geographic
area of the Channel Islands is used by the UN (M49), but not by ISO 3166-1,
and not by CLDR.  Thus, we are missing name translations for code 830 and
need to patch them in to avoid having a `nil` name value.

(Update CHANGELOG.md)