Improve Copy-on-Write friendliness
Lazily loading data files is an anti-pattern:

```ruby
module Currencies
  extend self

  def currency_codes
    @currency_codes ||= YAML.safe_load_file(CURRENCY_CODES_FILE_PATH)
  end
end
```

Here if `currency_codes` is never called during boot, it means that
it will be loaded by the first request that needs it, meaning:

  - A slower first call, responsible for "shark fin" looking latency graph around deploys.
  - The same data being loaded up to 32 times instead of once per pod (in core).
  - The same data using 32 times as much memory because it's not in CoW regions.
  - Will cause unecessary work for the GC because it won't immediately be in the old generation.
  - Will cause the memory of pods to grow over time looking like a leak.

If such data will be loaded anyway, it's much better to load it upfront
and essentially make it a boot defined constant.

Additionally, for this kind of data, it's best to pass `freeze: true`, to
`YAML.load_file` so it can deduplicate strings.

Here I fixed the most obvious occurences of that problem, but there
are a few left, like `@currencies_cache` and `@locales_cache`.