Stop reading paths files at load time
Trying to read these paths files any time this code is loaded breaks the cldr rake tasks if the paths files
don't exist (which may be the case if we've intentionally deleted them before regenerating them).

Move responsibility for reading the files into configure_i18n, which is what actually needs to use
the paths for loading i18n files.